Switch to adapter not composable
